+++ Myo Mapper/Source/emg.h
@@ -28,10 +28,10 @@ public:
 
     void paint (Graphics& g) override;
     void resized() override;
-    void setValues(std::array<int, 8> &EMG);
+    void setValues(std::array<int8_t, 8> &EMG);
     void setOSCPort (int Port);
     void setOSChostAddress(String HostAddress);
-    std::array<int, 8> getEmg();
+    std::array<int8_t, 8> getEmg();
     float getMav();
     
 private:
@@ -41,7 +41,7 @@ private:
     EmgS emgS;
     int oscPort;
     String hostAddress;
-    std::array<int, 8> EMGs;
+    std::array<int8_t, 8> EMGs;
     
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR (Emg)
 };
